Add an exclude for mu-plugins in dandelion.yml example
WPEngine has a number of plugins it maintains as an integration with its
own platform. These are write protected and will result in an error if
you don't exclude them from deployment. To help prevent this from
tripping people up in the future, I'm adding the neccesary exclusion
right in the example dandelion.yml included in the doc strings of the CI
configuration.